// src/types/profile-info.ts
export interface ProfileInfoInterface {
  birthDat?: string;
  profileCompletionPercentage?: number;
  userImage?: [];
  email?: string;
  phoneNumber?: string;
  lName?: string | null;
  fName?: string | null;
  userAbout?: string | null;
  linkdinProfile?: string | null;
  telegramLink?: string | null;
  receiveMessageEvent?: boolean;
  homeAdderess?: string | null;
  nationalCode?: string | null;
  gender?: boolean;
  birthDay?: string;
  latitude?: string | null;
  longitude?: string | null;
}